Fixed bug: Incorrected class hierachy lead to wrong logic
Former class hierachy:
Object->AbstractBaseEntity->MultiTenancyEntity->StandardObject
In former design, all entities are designed to extend StandardObjects.
But when we queried a User to get his tenantId, the query needs we
offering the tenantId, the logic doesn't make sense. So we decide to
make User, Group, and Role to non-multitenant.
The new class hierachy is:
Object->AbstractBaseEntity->StandardObject->MultiTenancyEntity